pyproject.toml
[project]
name = "s8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.0",
    "aiohttp-cors>=0.7.0",
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.7.4",
    "google-api-python-client>=2.108.0",
    "google-auth>=2.25.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.0.1.dev0",
    "mcp[cli]>=1.6.0",
    "ngrok>=1.4.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "pydantic>=2.0.0",
    "pymupdf4llm>=0.0.4",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "rich>=14.0.0",
    "tqdm>=4.64.0",
    "trafilatura[all]>=1.6.0",
]
